+++ src/main/java/cart/JwpCartApplication.java
@@ -3,10 +3,7 @@ package cart;
 import cart.domain.*;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.stereotype.Component;
-
-import java.util.Arrays;
 
 @SpringBootApplication
 public class JwpCartApplication {
